1.8.8 - Product Delete Check
Added a function comment and check to make sure a product exists before deleting it. This isn't important now due to the context of where it is called, however it would be important in the future if the application was scaled and this function where to be called out of context.

+++ app.py
@@ -215,9 +215,18 @@ def add_product(new_product):
 
 
 def delete_product(product_id):
+    """
+    Deletes a product from the database based on its product_id.
+
+    :param product_id: Integer representing the unique identifier of the product to be deleted.
+    :return: None
+    """
     the_product = session.query(Product).filter(Product.product_id==product_id).first()
-    session.delete(the_product)
-    session.commit()
+    if the_product:
+        session.delete(the_product)
+        session.commit()
+    else:
+        print(f"Product with ID {product_id} not found in the database.")
 
 
 def get_product_by_id():
